organize app runners and some more doc
